Tweak: Move device buttons above BlockCard
