+++ src/extend/inspector-control/controls/responsive-tabs/editor.scss
