+++ src/hoc/index.js
@@ -4,7 +4,6 @@ import withButtonLegacyMigration from './withButtonLegacyMigration';
 import withContainerLegacyMigration from './withContainerLegacyMigration';
 import withButtonContainerLegacyMigration from './withButtonContainerLegacyMigration';
 import './withDocumentation';
-import './withResponsiveTabs';
 
 export {
 	withUniqueId,
